fix: move ApplePay into apple-pay folder
